feat: cria AutorRepository e LivroRepository

+++ src/main/java/io/github/emanuelscapim/libraryapi/model/Livro.java
@@ -4,8 +4,8 @@ import io.github.emanuelscapim.libraryapi.model.enums.GeneroLivro;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -34,7 +34,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18,scale = 2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
